Removed typescript helper
